+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/handler/FormatterHandlerImpl.java
@@ -1,6 +1,7 @@
 package com.dwarfeng.settingrepo.impl.handler;
 
 import com.dwarfeng.settingrepo.stack.bean.entity.SettingCategory;
+import com.dwarfeng.settingrepo.stack.exception.FormatterException;
 import com.dwarfeng.settingrepo.stack.exception.FormatterMakeException;
 import com.dwarfeng.settingrepo.stack.exception.UnsupportedFormatterTypeException;
 import com.dwarfeng.settingrepo.stack.handler.Formatter;
@@ -22,22 +23,26 @@ public class FormatterHandlerImpl implements FormatterHandler {
     private List<FormatterMaker> formatterMakers = new ArrayList<>();
 
     @Override
-    public Formatter make(SettingCategory settingCategory)
-            throws UnsupportedFormatterTypeException, FormatterMakeException {
+    public Formatter make(SettingCategory settingCategory) throws FormatterException {
         try {
             // 生成触发器。
             LOGGER.debug("通过格式化器信息构建新的的格式化器...");
             FormatterMaker formatterMaker = formatterMakers.stream()
                     .filter(maker -> maker.supportType(settingCategory.getFormatterType())).findFirst()
                     .orElseThrow(() -> new UnsupportedFormatterTypeException(settingCategory.getFormatterType()));
-            Formatter formatter = formatterMaker.makeFormatter(settingCategory);
+            Formatter formatter;
+            try {
+                formatter = formatterMaker.makeFormatter(settingCategory);
+            } catch (Exception e) {
+                throw new FormatterMakeException(e);
+            }
             LOGGER.debug("格式化器构建成功!");
             LOGGER.debug("格式化器: " + formatter);
             return formatter;
-        } catch (UnsupportedFormatterTypeException e) {
+        } catch (FormatterException e) {
             throw e;
         } catch (Exception e) {
-            throw new FormatterMakeException(e);
+            throw new FormatterException(e);
         }
     }
 }