+++ settingrepo-impl/src/test/resources/groovy/TestFormatterProcessor.groovy
@@ -3,19 +3,20 @@ import com.dwarfeng.settingrepo.stack.bean.entity.SettingCategory
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey
 
 /**
- * 示例格式化器。
- *
- * <p>
- * 将输入的所有参数的 {@link Object#toString()} 返回的字符串简单的拼接在一起，形成的字符串作为主键的字符串返回。
+ * 测试用格式化器。
  */
 @SuppressWarnings("GrPackage")
-class ExampleFormatterProcessor implements GroovyFormatterRegistry.Processor {
+class TestFormatterProcessor implements GroovyFormatterRegistry.Processor {
 
     @Override
-    StringIdKey format(SettingCategory settingCategory, Object[] args) throws Exception {
+    StringIdKey format(SettingCategory settingCategory, String[] args) throws Exception {
         StringBuilder sb = new StringBuilder()
-        for (Object arg : args) {
+        for (int i = 0; i < args.length; i++) {
+            String arg = args[i]
             sb.append(arg)
+            if (i < args.length - 1) {
+                sb.append('.')
+            }
         }
         return new StringIdKey(sb.toString())
     }