+++ settingrepo-stack/src/main/java/com/dwarfeng/settingrepo/stack/handler/Formatter.java
@@ -1,6 +1,6 @@
 package com.dwarfeng.settingrepo.stack.handler;
 
-import com.dwarfeng.settingrepo.stack.exception.FormatterExecutionException;
+import com.dwarfeng.settingrepo.stack.exception.FormatterException;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.handler.Handler;
 
@@ -17,7 +17,7 @@ public interface Formatter extends Handler {
      *
      * @param args 格式化所需的参数组成的数组。
      * @return 参数格式化后形成的配置节点主键。
-     * @throws FormatterExecutionException 格式化过程中出现异常。
+     * @throws FormatterException 格式化器异常。
      */
-    StringIdKey format(Object[] args) throws FormatterExecutionException;
+    StringIdKey format(String[] args) throws FormatterException;
 }