+++ settingrepo-stack/src/main/java/com/dwarfeng/settingrepo/stack/handler/FormatterHandler.java
@@ -1,8 +1,7 @@
 package com.dwarfeng.settingrepo.stack.handler;
 
 import com.dwarfeng.settingrepo.stack.bean.entity.SettingCategory;
-import com.dwarfeng.settingrepo.stack.exception.FormatterMakeException;
-import com.dwarfeng.settingrepo.stack.exception.UnsupportedFormatterTypeException;
+import com.dwarfeng.settingrepo.stack.exception.FormatterException;
 
 /**
  * 格式化器处理器。
@@ -17,8 +16,7 @@ public interface FormatterHandler {
      *
      * @param settingCategory 指定的设置类别。
      * @return 指定的设置类别构造的格式化器。
-     * @throws UnsupportedFormatterTypeException 不支持的格式化器类型。
-     * @throws FormatterMakeException            格式化器构造失败。
+     * @throws FormatterException 格式化器异常。
      */
-    Formatter make(SettingCategory settingCategory) throws UnsupportedFormatterTypeException, FormatterMakeException;
+    Formatter make(SettingCategory settingCategory) throws FormatterException;
 }